Moving the methods for replaying runs into separate file.
