+++ src/testup/ui.rb
@@ -1,9 +1,11 @@
 #-------------------------------------------------------------------------------
 #
-# Copyright 2013-2014 Trimble Navigation Ltd.
+# Copyright 2013-2016 Trimble Inc.
 # License: The MIT License (MIT)
 #
 #-------------------------------------------------------------------------------
+
+require 'testup/runs'
 
 
 module TestUp
@@ -49,14 +51,14 @@ module TestUp
     cmd_seed = cmd
 
     cmd = UI::Command.new('Re-run Saved Run') {
-      self.rerun_fixed_run
+      self::Runs.rerun_current
     }
     cmd.tooltip = 'Re-run a Saved Run from a previously picked .run file.'
     cmd.status_bar_text = 'Re-run a Saved Run from a previously picked .run file.'
     cmd.small_icon = File.join(PATH_IMAGES, 'rerun.png')
     cmd.large_icon = File.join(PATH_IMAGES, 'rerun.png')
     cmd.set_validation_proc {
-      self.current_run ? MF_ENABLED : MF_GRAYED
+      self::Runs.current ? MF_ENABLED : MF_GRAYED
     } if defined?(Sketchup)
     cmd_rerun_saved = cmd
 
@@ -113,7 +115,7 @@ module TestUp
     cmd.tooltip = 'Run Layout Tests'
     cmd.small_icon = File.join(PATH_IMAGES, 'layout-16.png')
     cmd.large_icon = File.join(PATH_IMAGES, 'layout-24.png')
-    cmd_run_layout_tests = cmd
+    #cmd_run_layout_tests = cmd
 
     # Menus
     if defined?(Sketchup)
@@ -123,15 +125,15 @@ module TestUp
       menu.add_item(cmd_seed)
       menu.add_separator
       sub_menu = menu.add_submenu('Saved Runs')
-      menu_id = sub_menu.add_item('Set Replay-Run') { self.set_fixed_run }
+      menu_id = sub_menu.add_item('Set Replay-Run') { self::Runs.set_current }
       sub_menu.set_validation_proc(menu_id) {
-        self.current_run ? MF_CHECKED : MF_ENABLED
+        self::Runs.current ? MF_CHECKED : MF_ENABLED
       }
       sub_menu.add_separator
       sub_menu.add_item(cmd_rerun_saved)
       sub_menu.add_separator
-      sub_menu.add_item('Add Run...') { self.add_run }
-      sub_menu.add_item('Remove Run...') { self.remove_run }
+      sub_menu.add_item('Add Run...') { self::Runs.add }
+      sub_menu.add_item('Remove Run...') { self::Runs.remove }
       menu.add_separator
       menu.add_item(cmd_open_logs)
       menu.add_item(cmd_toggle_run_tests_in_console)
@@ -192,14 +194,13 @@ module TestUp
 
     def self.select_directory_fallback(options)
       unless RUBY_PLATFORM =~ /mswin|mingw/
-        warn "select_directory_fallback not implemented for this platform"
+        warn 'select_directory_fallback not implemented for this platform'
         return nil
       end
 
       require 'win32ole'
-      default_path = File.join(ENV['HOME'], 'Desktop').gsub('/', '\\')
 
-      message = ""
+      message = ''
       select_multiple = false
       if options
         message = options[:message] || message
@@ -211,9 +212,11 @@ module TestUp
       dialog_options = BIF_RETURNONLYFSDIRS | BIF_NEWDIALOGSTYLE | BIF_EDITBOX
 
       # http://msdn.microsoft.com/en-us/library/windows/desktop/bb774065(v=vs.85).aspx
+      # noinspection RubyResolve
       objFolder = objShell.BrowseForFolder(parent_window, message, dialog_options)
 
       return nil if objFolder.nil?
+      # noinspection RubyResolve
       path = objFolder.Self.Path
       unless File.exist?(path)
         UI.messagebox("Unable to handle '#{path}'.")
